+++ src/main/java/shop/tryit/web/order/OrderController.java
@@ -16,9 +16,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import shop.tryit.domain.common.Pages;
-import shop.tryit.domain.order.OrderFacade;
 import shop.tryit.domain.order.dto.OrderFormDto;
 import shop.tryit.domain.order.dto.OrderSearchDto;
+import shop.tryit.domain.order.service.OrderFacade;
 
 @Slf4j
 @Controller
